requirements.txt
Django==4.1.3
django_rest_framework_mongoengine==3.4.1
djangorestframework==3.14.0
mongoengine==0.24.2
pymongo==4.3.3
python_dateutil==2.8.2
